Refactoring to use the pages router. Fixing issues
