+++ components/proposal/details.tsx
