+++ components/vote/votes-section.tsx
