+++ components/vote/voting-modal.tsx
