+++ package.json
@@ -9,7 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@aragon/ods": "^1.0.5",
+    "@aragon/ods": "1.0.7",
     "@shazow/whatsabi": "^0.9.1",
     "@web3modal/wagmi": "^3.5.6",
     "dayjs": "^1.11.10",
